+++ internal/conversations/test_main_test.go
@@ -6,7 +6,11 @@ import (
 	"testing"
 
 	"github.com/teanode/teanode/internal/configs"
+	"github.com/teanode/teanode/internal/store"
+	storefs "github.com/teanode/teanode/internal/store/fs"
 )
+
+var openedStoreForTests store.Store
 
 func TestMain(m *testing.M) {
 	directory, err := os.MkdirTemp("", "teanode-conversations-tests-")
@@ -15,7 +19,16 @@ func TestMain(m *testing.M) {
 		os.Exit(1)
 	}
 	configs.SetDirectory(directory)
+	openedStore, openError := storefs.Open(storefs.Options{DataDirectory: directory})
+	if openError != nil {
+		fmt.Fprintf(os.Stderr, "store open failed: %v\n", openError)
+		os.Exit(1)
+	}
+	openedStoreForTests = openedStore
 	exitCode := m.Run()
+	if closeError := openedStore.Close(); closeError != nil {
+		fmt.Fprintf(os.Stderr, "store close failed: %v\n", closeError)
+	}
 	if removeErr := os.RemoveAll(directory); removeErr != nil {
 		fmt.Fprintf(os.Stderr, "RemoveAll failed: %v\n", removeErr)
 	}
